refactor: apply barriers list for vanilla
